Move ValueHandler to separate files
