+++ src/utils/confighandler.cpp
@@ -5,6 +5,7 @@
 #include "src/tools/capturetool.h"
 #include "src/utils/configshortcuts.h"
 #include "systemnotification.h"
+#include "valuehandler.h"
 #include <QCoreApplication>
 #include <QDebug>
 #include <QDir>
@@ -21,37 +22,6 @@
 #endif
 
 // HELPER FUNCTIONS
-
-QList<int> fromButtonToInt(const QList<CaptureToolButton::ButtonType>& l)
-{
-    QList<int> buttons;
-    for (auto const i : l)
-        buttons << static_cast<int>(i);
-    return buttons;
-}
-
-QList<CaptureToolButton::ButtonType> fromIntToButton(const QList<int>& l)
-{
-    QList<CaptureToolButton::ButtonType> buttons;
-    for (auto const i : l)
-        buttons << static_cast<CaptureToolButton::ButtonType>(i);
-    return buttons;
-}
-
-bool normalizeButtons(QList<int>& buttons)
-{
-    QList<int> listTypesInt =
-      fromButtonToInt(CaptureToolButton::getIterableButtonTypes());
-
-    bool hasChanged = false;
-    for (int i = 0; i < buttons.size(); i++) {
-        if (!listTypesInt.contains(buttons.at(i))) {
-            buttons.removeAt(i);
-            hasChanged = true;
-        }
-    }
-    return hasChanged;
-}
 
 bool verifyLaunchFile()
 {
@@ -73,300 +43,6 @@ bool verifyLaunchFile()
 }
 
 // VALUE HANDLING
-
-/**
- * Handles the value of a configuration option.
- *
- * Each configuration option should usually be handled in three different ways:
- * - have its value checked for errors (type, format, etc.)
- * - have its value (that was taken from the config file) adapted for proper use
- * - provided a fallback value in case: the config does not explicitly specify
- *   it, or the config contains an error and is globally falling back to
- *   defaults
- *
- * Subclass this class to handle custom value types.
- *
- * If you wish to handle simple value types (those supported by QVariant) you
- * should use `SimpleValueHandler`. Note that you can't use that class if the
- * value has some custom constraints on it. TODO
- *
- * @note You will only need to override `get` if you have to change the value
- * that was read from the config file. If you are fine with the value as long as
- * it is error-free, you don't have to override it.
- *
- * @note Keep in mind that you will probably want `check` to return `true` for
- * invalid QVariant's (option not found in config file).
- *
- */
-class ValueHandler
-{
-public:
-    virtual bool check(const QVariant& val) = 0;
-    virtual QVariant value(const QVariant& val)
-    {
-        if (!val.isValid() || !check(val)) {
-            return fallback();
-        } else {
-            return process(val);
-        }
-    }
-    virtual QVariant fallback() { return QVariant(); };
-    virtual QVariant representation(const QVariant& val)
-    {
-        return val.toString();
-    }
-
-protected:
-    virtual QVariant process(const QVariant& val) { return val; }
-};
-
-// CUSTOM TYPE HANDLERS
-
-class Bool : public ValueHandler
-{
-public:
-    Bool(bool def)
-      : m_def(def)
-    {}
-    bool check(const QVariant& val) override
-    {
-        QString str = val.toString();
-        if (str != "true" && str != "false") {
-            return false;
-        }
-        return true;
-    }
-    QVariant fallback() override { return m_def; }
-
-private:
-    bool m_def;
-};
-
-class String : public ValueHandler
-{
-public:
-    String(const QString& def)
-      : m_def(def)
-    {}
-    bool check(const QVariant&) override { return true; }
-    QVariant fallback() override { return m_def; }
-
-private:
-    QString m_def;
-};
-
-class Color : public ValueHandler
-{
-public:
-    Color(const QColor& def)
-      : m_def(def)
-    {}
-    bool check(const QVariant& val) override
-    {
-        return QColor::isValidColor(val.toString());
-    }
-    QVariant fallback() override { return m_def; }
-    QVariant representation(const QVariant& val) override
-    {
-        return QString(val.value<QColor>().name());
-    }
-
-private:
-    QColor m_def;
-};
-
-class BoundedInt : public ValueHandler
-{
-public:
-    BoundedInt(int min, int max, int def)
-      : m_min(min)
-      , m_max(max)
-      , m_def(def)
-    {}
-
-    bool check(const QVariant& val) override
-    {
-        QString str = val.toString();
-        bool conversionOk;
-        int num = str.toInt(&conversionOk);
-        return conversionOk && (m_max < m_min || num <= m_max);
-    }
-    virtual QVariant fallback() override { return m_def; };
-
-private:
-    int m_min, m_max, m_def;
-};
-
-class LowerBoundedInt : public ValueHandler
-{
-public:
-    LowerBoundedInt(int min, int def)
-      : m_min(min)
-      , m_def(def)
-    {}
-    bool check(const QVariant& val) override
-    {
-        QString str = val.toString();
-        bool conversionOk;
-        int num = str.toInt(&conversionOk);
-        return conversionOk && num >= m_min;
-    };
-    virtual QVariant fallback() override { return m_def; }
-
-private:
-    int m_min, m_def;
-};
-
-class KeySequence : public ValueHandler
-{
-public:
-    KeySequence(const QString& shortcutName)
-      : m_shortcutName(shortcutName)
-    {}
-    bool check(const QVariant& val) override
-    {
-        // TODO
-        return true;
-    }
-    QVariant fallback() override
-    {
-        return ConfigShortcuts()
-          .captureShortcutDefault(m_shortcutName)
-          .toString();
-    }
-
-private:
-    QString m_shortcutName;
-};
-
-class ExistingDir : public ValueHandler
-{
-    bool check(const QVariant& val) override
-    {
-        if (!val.canConvert(QVariant::String) || val.toString().isEmpty()) {
-            return false;
-        }
-        QFileInfo info(val.toString());
-        return info.isDir() && info.exists();
-    }
-
-    QVariant fallback() override
-    {
-        QString path =
-          QStandardPaths::writableLocation(QStandardPaths::PicturesLocation);
-        if (!QFileInfo(path).isDir()) {
-            // TODO can we rely on the fact that home exists?
-            path =
-              QStandardPaths::writableLocation(QStandardPaths::HomeLocation);
-        }
-        return path;
-    }
-};
-
-class FilenamePattern : public ValueHandler
-{
-    bool check(const QVariant&) override { return true; }
-    QVariant fallback() override
-    {
-        return ConfigHandler().filenamePatternDefault();
-    }
-};
-
-class ButtonList : public ValueHandler
-{
-    using BList = QList<CaptureToolButton::ButtonType>;
-    bool check(const QVariant& val) override
-    {
-        auto allButtons = CaptureToolButton::getIterableButtonTypes();
-        using CTB = CaptureToolButton;
-        for (int btn : val.value<QList<int>>()) {
-            if (!allButtons.contains(static_cast<CTB::ButtonType>(btn))) {
-                return false;
-            }
-        }
-        return true;
-    }
-    QVariant value(const QVariant& val) override
-    {
-        // Get unsorted button list
-        BList buttons = ValueHandler::value(val).value<BList>();
-
-        using BT = CaptureToolButton::ButtonType;
-        std::sort(buttons.begin(), buttons.end(), [](BT a, BT b) {
-            return CaptureToolButton::getPriorityByButton(a) <
-                   CaptureToolButton::getPriorityByButton(b);
-        });
-        return QVariant::fromValue(buttons);
-    }
-    QVariant process(const QVariant& val) override
-    {
-        QList<int> intButtons = val.value<QList<int>>();
-        auto buttons = fromIntToButton(intButtons);
-        return QVariant::fromValue(buttons);
-    }
-    QVariant fallback() override
-    {
-        auto buttons = CaptureToolButton::getIterableButtonTypes();
-        buttons.removeOne(CaptureToolButton::TYPE_SIZEDECREASE);
-        buttons.removeOne(CaptureToolButton::TYPE_SIZEINCREASE);
-        // TODO: remove toList in v1.0
-        return QVariant::fromValue(buttons);
-    }
-    QVariant representation(const QVariant& val) override
-    {
-        auto intList = fromButtonToInt(val.value<BList>());
-        normalizeButtons(intList);
-        return QVariant::fromValue(intList);
-    }
-};
-
-class UserColors : public ValueHandler
-{
-    bool check(const QVariant& val) override
-    {
-        if (!val.isValid()) {
-            return true;
-        }
-        if (!val.canConvert(QVariant::StringList)) {
-            return false;
-        }
-        for (const QString& str : val.toStringList()) {
-            if (!QColor::isValidColor(str)) {
-                return false;
-            }
-        }
-        return true;
-    }
-    QVariant process(const QVariant& val) override
-    {
-        QStringList strColors = val.toStringList();
-        if (strColors.isEmpty()) {
-            return fallback();
-        }
-
-        QVector<QColor> colors;
-        colors.reserve(strColors.size());
-
-        for (const QString& str : strColors) {
-            colors.append(QColor(str));
-        }
-
-        return QVariant::fromValue(colors);
-    }
-    QVariant fallback() override
-    {
-        return QVariant::fromValue(QVector<QColor>{ Qt::darkRed,
-                                                    Qt::red,
-                                                    Qt::yellow,
-                                                    Qt::green,
-                                                    Qt::darkGreen,
-                                                    Qt::cyan,
-                                                    Qt::blue,
-                                                    Qt::magenta,
-                                                    Qt::darkMagenta,
-                                                    QColor() });
-    }
-};
 
 /**
  * Use this to declare a setting with a type that is either unrecognized by
@@ -578,7 +254,7 @@ QString ConfigHandler::saveAsFileExtension()
 void ConfigHandler::setAllTheButtons()
 {
     QList<int> buttons =
-      fromButtonToInt(CaptureToolButton::getIterableButtonTypes());
+      ButtonList::toIntList(CaptureToolButton::getIterableButtonTypes());
     setValue(QStringLiteral("buttons"), QVariant::fromValue(buttons));
 }
